Extract circular_indexing into a separate header
